Update save state and result control by Hiep

+++ Fighter/Assets/Scripts/Save/SaveState.cs
@@ -65,7 +65,7 @@ public class SaveState {
 
 	public int[] iconChar = new int[8]  { 0, 0, 0, 0, 0, 0, 0, 0};
 
-	public int score = 0;
+	public float score = 0;
 
 	public bool firstInGame = false;
